+++ CMakeLists.txt
@@ -63,7 +63,7 @@ add_subdirectory(libcontract)
 add_subdirectory(libsinglepoint)
 add_subdirectory(libpbftseal)
 add_subdirectory(libraftseal)
-add_subdirectory(libpailler)
+add_subdirectory(libpaillier)
 add_subdirectory(libethcore)
 add_subdirectory(libethereum)
 add_subdirectory(libevm)
